refactor

+++ controllers/book.js
@@ -3,22 +3,26 @@ const Book = require("../models/book");
 // HTTP errors are being handled with express-async-errors
 // therefore no need to write catch
 
-// Get all books
+// Get all
 const getAllBooks = async (req, res) => {
-  const users = await Book.find({});
-  res.status(200).json({ users: users });
+  const books = await Book.find({});
+  res.status(200).json({ books: books });
 };
 
-// Get a book by id
+// Get
 const getBook = async (req, res) => {
   const book = await Book.findById(req.params.id);
-  res.json(book);
+
+  if (!book) {
+    res.status(400).send(`No such book with the ID: ${req.params.id} `);
+  } else {
+    res.status(200).json(book);
+  }
 };
 
-// Insert a new book
-const addBook = async (req, res) => {
-  const { body } = req;
-  const { title, author, language, likes } = body;
+// Create
+const createBook = async (req, res) => {
+  const { title, author, language, likes } = req.body;
   const newBookObj = new Book({
     title,
     author,
@@ -30,21 +34,33 @@ const addBook = async (req, res) => {
   res.status(201).json(newBook);
 };
 
-// Update a book
+// Update
 const updateBook = async (req, res) => {
-  const { id } = req.params;
+  const bookID = req.params.id;
+  const book = Book.findById(bookID);
   const newBookObj = req.body;
 
-  const updatedBook = await Book.findByIdAndUpdate(id, newBookObj, {
-    new: true,
-  });
-  res.json(updatedBook);
+  if (!book) {
+    res.status(400).send(`No such book with the ID: ${bookID}`);
+  } else {
+    const updatedBook = await Book.findByIdAndUpdate(bookID, newBookObj, {
+      new: true,
+    });
+    res.json(updatedBook);
+  }
 };
 
-// Delete book
+// Delete
 const deleteBook = async (req, res) => {
-  const book = await Book.findByIdAndDelete(req.params.id);
-  res.status(204);
+  const bookID = req.params.id;
+  const book = await Book.findById(bookID);
+
+  if (!book) {
+    res.status(400).send(`No such book with the ID: ${bookID}`);
+  } else {
+    await Book.findByIdAndDelete(req.params.id);
+    res.status(204);
+  }
 };
 
-module.exports = { getAllBooks, addBook, updateBook, getBook, deleteBook };
+module.exports = { getAllBooks, createBook, updateBook, getBook, deleteBook };
